refactor(profile): 抽取插件启停/卸载管理到 profile/plugin-manager.js
main.js 的插件管理状态簇(18 个函数 + 局部 yaml 方言状态 + 图片粘贴,
约 370 行)迁出为独立模块:

- pluginManagerCollect/SetEnabled/SetRemoved/Uninstall/Restore 四个 IPC
  驱动的实现 + imagePasteSave + patch 读写助手,全部原样迁移
- 依赖注入:稳定函数/常量按引用;serverProc/restartingServer(可变 let)
  经 getServerProc/getRestartingServer 调用期取值
- 循环依赖处理:pluginManager 需要 syncCompanionPlugins(companionSync
  在其后初始化),用调用时包装 () => syncCompanionPlugins() 传入,
  运行时解引用无 TDZ;companionSync 需要的 removedPluginIds 由本模块
  解构提供(接线顺序:pluginManager → companionSync)
- __dirname 上溯一层修正(profile/ → ../assets/plugins)
- main.js 净减约 370 行(4557 → 3327,累计 -1230)
- electron-builder.yml / check-syntax.js 清单同步
- 新增 8 个组件测试(settings 往返/patch 兜底解析/启停写盘/收集/
  卸载流程/核心保护/图片粘贴边界)
- 测试发现并修正两处测试自身缺陷:核心拦截在 SetRemoved 而非
  Uninstall;base64 20MB 恰好等于 15MB 上限边界

测试: 464 pass / 1 skip / 0 fail,git diff --check 通过
