+++ dsh-desktop/scripts/check-syntax.js
@@ -36,6 +36,7 @@ const entryFiles = [
   'ipc/register-ipc.js',
   'profile/runtime-patches.js',
   'profile/companion-sync.js',
+  'profile/plugin-manager.js',
 ];
 
 // 匹配「async/await 关键字与紧随其后的 function 声明之间被空行/注释行拆开」：
